requirements.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
html2text==2018.1.9
idna==2.8
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.16.4
protobuf==3.9.0
pyparsing==2.4.1.1
python-dateutil==2.8.0
python-slugify==3.0.2
PyYAML==5.1.1
requests==2.22.0
scipy==1.3.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
text-unidecode==1.2
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
